Rename directoryToObject helper
