+++ lib/directory-to-object.js
